made process start and stop robust
